Changed blockCount parameter in eth_feeHistory to long. (#3670)

+++ src/Nethermind/Nethermind.JsonRpc/Modules/Eth/FeeHistory/IFeeHistoryOracle.cs
@@ -21,7 +21,7 @@ namespace Nethermind.JsonRpc.Modules.Eth.FeeHistory
 {
     public interface IFeeHistoryOracle
     {
-        ResultWrapper<FeeHistoryResults> GetFeeHistory(int blockCount, BlockParameter newestBlock, double[]? rewardPercentiles);
+        ResultWrapper<FeeHistoryResults> GetFeeHistory(long blockCount, BlockParameter newestBlock, double[]? rewardPercentiles);
         
     }
 }
